requirements.txt
six==1.10.0
websocket-client==0.37.0
